pyproject.toml
[project]
name = "astronomia-ui-backend"
version = "0.1.0"
description = "BFF para astronomIA UI. Conecta el frontend con el backend de análisis (o con n8n cuando exista)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.32.0",
  "httpx>=0.27.0",
  "pydantic>=2.9.0",
  "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
]
